fixing for makefile

+++ src/ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: kkweon <kkweon@student.codam.nl>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/07 16:51:50 by kkweon            #+#    #+#             */
-/*   Updated: 2025/10/13 11:13:47 by kkweon           ###   ########.fr       */
+/*   Updated: 2025/10/23 14:52:56 by kkweon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <libft.h>
+
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int i;
+	size_t i;
 	char *s1_c;
 	char *s2_c;
